requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.8
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
httpie==3.1.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
MarkupPy==1.14
multidict==6.0.2
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
Pygments==2.11.2
PySocks==1.7.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
